backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Libraries
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tensorflow==2.18.0
transformers==4.46.3
opencv-python==4.10.0.84
librosa==0.10.2
scikit-learn==1.5.2
numpy==1.26.4
pillow==10.4.0

# Data Processing
pandas==2.2.3
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Handling
aiofiles==24.1.0
python-magic==0.4.27

# Utilities
pydantic==2.10.3
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Monitoring
prometheus-client==0.21.1